Refactor: 코멘트 반영 #198
