+++ src/main/java/com/nhnacademy/marketgg/server/repository/customerservicecomment/CustomerServiceCommentRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.marketgg.server.repository.cart.customerservicecomment;
+package com.nhnacademy.marketgg.server.repository.customerservicecomment;
 
 import org.springframework.data.repository.NoRepositoryBean;
