+++ src/main/java/com/nhnacademy/marketgg/server/repository/customerservicecomment/CustomerServiceCommentRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.marketgg.server.repository.cart.customerservicecomment;
+package com.nhnacademy.marketgg.server.repository.customerservicecomment;
 
 import com.nhnacademy.marketgg.server.entity.CustomerServiceComment;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
