+++ src/test/java/com/nhnacademy/marketgg/server/service/impl/DefaultPostServiceTest.java
@@ -32,7 +32,7 @@ import com.nhnacademy.marketgg.server.entity.Category;
 import com.nhnacademy.marketgg.server.entity.CustomerServicePost;
 import com.nhnacademy.marketgg.server.entity.Member;
 import com.nhnacademy.marketgg.server.repository.category.CategoryRepository;
-import com.nhnacademy.marketgg.server.repository.cart.customerservicecomment.CustomerServiceCommentRepository;
+import com.nhnacademy.marketgg.server.repository.customerservicecomment.CustomerServiceCommentRepository;
 import com.nhnacademy.marketgg.server.repository.customerservicepost.CustomerServicePostRepository;
 import com.nhnacademy.marketgg.server.repository.member.MemberRepository;
 import com.nhnacademy.marketgg.server.service.post.DefaultPostService;
